add module, add text, add responsiveness
