+++ index.html
@@ -80,6 +80,7 @@
                 </div>
                 <div id="step3" class="hidden">
                     <h2 class="title">Soorten laatste contact met de politie per groep</h2>
+                    <p class="first-sentence">Over de <span class="yellow">693</span> respondenten die wel contact hebben gehad met de politie, is dit de onderverdeling of de politie met de respondenent opnam, of andersom.</p>
                     <div id="bar__container">
                         <svg width="700" height="400" class = "stack" id="stacked-bar"></svg>
                         <div class="step2-legend">
@@ -90,7 +91,6 @@
                 </div>
                 <div id="step4" class="hidden">
                     <h2 class="title">Gevolgen laatste contact</h2>
-
                     <h4>Filter op gevolg: </h4>
                     <select id="selectButton">
                             <option value="niets">Niets</option>
@@ -108,12 +108,15 @@
                 <div id="step5" class="hidden">
                     <h2 class="title">Hoe verliep het laatste contact met de politie?</h2>
                     <div class="chart-with-legend__container">
-                        <svg width="960" height="450" class = "groupedBars"></svg>
+                        <svg width="700" height="400" class = "groupedBars"></svg>
                         <ul class="legend">
                             <h2 class="legend-title">Legenda</h2>
                             <li class="NL">Nederlandse respondenten</li>
                             <li class="WM">Westerse respondenten</li>
                             <li class="NWM">niet-Westerse respondenten</li>
+                            <br>
+                            <li>1.0 = Helemaal mee oneens</li>
+                            <li>5.0 = Helemaal mee eens</li>
                         </ul>
                     </div>
                 </div>
