+++ public/global.css
@@ -147,6 +147,7 @@ h2 {
 
 .yellow {
     color: #FFF33D;
+    font-family: Rubik-bold;
 }
 
 .blue {
@@ -167,14 +168,10 @@ h2 {
 }
 
 /* Pie chart */
-.pie {
-    margin-left: 10em;
-}
-
 .chart-with-legend__container {
-    position: absolute;
-    right: 4em;
     display: flex;
+    position: absolute;
+    right: 3%;
 }
 
 .pie__container {
@@ -190,6 +187,7 @@ h2 {
 .first-sentence {
     margin-left: 3em;
     font-family: Roboto-light;
+    max-width: 24em;
 }
 
 .pie g path:hover {
@@ -205,14 +203,10 @@ h2 {
 }
 
 /* Bar chart */
-#stacked-bar {
-    /* margin-left: 3em; */
-    min-width: 290px;
-}
-
 #bar__container {
     display: flex;
     flex-direction: row;
+    
 }
 
 .stack g g rect:hover {
@@ -417,10 +411,14 @@ circle{
     width: 90%;
 }
 
-@media only screen and (max-width: 320px){
-    
+@media screen and (max-width: 320px){
 }
 
-@media only screen and (min-width: 600px){
-
+@media screen and (max-width: 600px){
+    #introduction__container {
+        width: 80%;
+    }
+    #top__container, #bottom__container {
+        width: 100%;
+    }
 }
